% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_est_moe}
\alias{fmt_est_moe}
\title{Format estimate and MOE columns in a gt table}
\usage{
fmt_est_moe(
  gt_object,
  cols = c("estimate", "moe"),
  est_col_label = "Est.",
  spanner = NULL,
  decimals = 0,
  use_seps = TRUE
)
}
\description{
Format estimate and MOE columns in a gt table
}
\keyword{internal}
