% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{fmt_perc_moe}
\alias{fmt_perc_moe}
\title{Format \% estimate and \% MOE columns in a gt table}
\usage{
fmt_perc_moe(
  gt_object,
  cols = c("perc_estimate", "moe_perc"),
  perc_col_label = "\% share",
  spanner = NULL,
  decimals = 0,
  use_seps = TRUE
)
}
\description{
Format \% estimate and \% MOE columns in a gt table
}
\keyword{internal}
