% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gt_census_source}
\alias{gt_census_source}
\title{Add a source note to a table}
\usage{
gt_census_source(
  gt_object,
  source = "Source: 2017-2021 ACS 5-year Estimates",
  tables = NULL
)
}
\description{
Add a source note to a table
}
\keyword{internal}
