% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_name_ext}
\alias{has_name_ext}
\title{Helper for checking name availability}
\usage{
has_name_ext(
  x,
  name = NULL,
  starts_with = NULL,
  ends_with = NULL,
  contains = NULL
)
}
\description{
Helper for checking name availability
}
\keyword{internal}
