% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{label_acs}
\alias{label_acs}
\title{Label acs data using table and column metadata from Census Reporter}
\usage{
label_acs(data, survey = "acs5", year = 2021, perc = TRUE)
}
\description{
Label acs data using table and column metadata from Census Reporter
}
\keyword{internal}
