% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_census_reporter_metadata}
\alias{load_census_reporter_metadata}
\title{Load metadata from Census Reporter}
\usage{
load_census_reporter_metadata(
  survey = "acs5",
  year = 2021,
  metadata = "table",
  ...
)
}
\arguments{
\item{survey}{Survey using options from \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}}

\item{year}{Year of sample to return.}

\item{metadata}{Type of metadata to return, "table" or "column"}
}
\description{
See \url{https://github.com/censusreporter/census-table-metadata}
}
\keyword{internal}
